refactor(general): :recycle: Add barrel file to all modules and fix imports
