+++ src/app.module.ts
@@ -1,19 +1,17 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 
-import { UserModule } from './modules';
-import { AuthModule } from './providers';
-import { CatalogsModule } from './catalogs/catalogs/catalogs.module';
-import { ModulesModule } from './catalogs/modules/modules.module';
-import { PermissionsModule } from './catalogs/permissions/permissions.module';
-import { RolesModule } from './catalogs/roles/roles.module';
-import { StatesModule } from './catalogs/states/states.module';
-import { TypeTestModule } from './catalogs/type-test/type-test.module';
-import { TypesModuleModule } from './catalogs/types-module/types-module.module';
-import { NodemailerModule } from './providers/nodemailer/nodemailer.module';
-import { EnterprisesModule } from './modules/enterprises/enterprises.module';
-import { FirebaseModule } from './providers/firebase/firebase.module';
-import { ContractsModule } from './modules/contracts/contracts.module';
+import {
+  CatalogsModule,
+  ModulesModule,
+  PermissionsModule,
+  RolesModule,
+  StatesModule,
+  TypesModuleModule,
+  TypeTestModule,
+} from './catalogs';
+import { ContractsModule, EnterprisesModule, UserModule } from './modules';
+import { AuthModule, FirebaseModule, NodemailerModule } from './providers';
 
 @Module({
   imports: [
