+++ src/catalogs/catalogs/index.ts
@@ -1,0 +1,3 @@
+export * from './catalogs.module';
+export * from './controllers';
+export * from './services';
