+++ src/catalogs/index.ts
@@ -1,3 +1,4 @@
+export * from './catalogs';
 export * from './cities';
 export * from './modules';
 export * from './permissions';
