+++ src/catalogs/type-test/index.ts
@@ -1,3 +1,3 @@
-export * from './helpers';
 export * from './interfaces';
+export * from './services';
 export * from './type-test.module';
