+++ src/modules/contracts/index.ts
@@ -1,0 +1,5 @@
+export * from './contracts.module';
+export * from './controllers';
+export * from './dtos';
+export * from './interfaces';
+export * from './services';
