+++ src/providers/index.ts
@@ -1,2 +1,4 @@
 export * from './auth';
 export * from './bcrypt';
+export * from './firebase';
+export * from './nodemailer';
